requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.8
et_xmlfile==2.0.0
idna==3.10
iniconfig==2.0.0
mysql-connector-python==9.1.0
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.4
pytest-cov==6.0.0
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
